refactor(trie): move auxiliary sparse trie types into modules
