+++ crates/trie/sparse/src/trie.rs
@@ -1,10 +1,13 @@
-use crate::blinded::{BlindedProvider, DefaultBlindedProvider, RevealedNode};
+use crate::{
+    blinded::{BlindedProvider, DefaultBlindedProvider, RevealedNode},
+    RemovedSparseNode, RlpNodeBuffers, RlpNodePathStackItem, RlpNodeStackItem, SparseNode,
+    SparseNodeType, SparseTrieUpdates,
+};
 use alloc::{
     borrow::Cow,
     boxed::Box,
     fmt,
     string::{String, ToString},
-    vec,
     vec::Vec,
 };
 use alloy_primitives::{
@@ -19,7 +22,6 @@ use reth_trie_common::{
     BranchNodeCompact, BranchNodeRef, ExtensionNodeRef, LeafNodeRef, Nibbles, RlpNode, TrieMask,
     TrieNode, CHILD_INDEX_RANGE, EMPTY_ROOT_HASH,
 };
-use smallvec::SmallVec;
 use tracing::trace;
 
 /// Struct for passing around branch node mask information.
@@ -1919,233 +1921,6 @@ impl<P: BlindedProvider> RevealedSparseTrie<P> {
         }
 
         Ok(())
-    }
-}
-
-/// Enum representing sparse trie node type.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum SparseNodeType {
-    /// Empty trie node.
-    Empty,
-    /// A placeholder that stores only the hash for a node that has not been fully revealed.
-    Hash,
-    /// Sparse leaf node.
-    Leaf,
-    /// Sparse extension node.
-    Extension {
-        /// A flag indicating whether the extension node should be stored in the database.
-        store_in_db_trie: Option<bool>,
-    },
-    /// Sparse branch node.
-    Branch {
-        /// A flag indicating whether the branch node should be stored in the database.
-        store_in_db_trie: Option<bool>,
-    },
-}
-
-impl SparseNodeType {
-    const fn is_hash(&self) -> bool {
-        matches!(self, Self::Hash)
-    }
-
-    const fn is_branch(&self) -> bool {
-        matches!(self, Self::Branch { .. })
-    }
-
-    const fn store_in_db_trie(&self) -> Option<bool> {
-        match *self {
-            Self::Extension { store_in_db_trie } | Self::Branch { store_in_db_trie } => {
-                store_in_db_trie
-            }
-            _ => None,
-        }
-    }
-}
-
-/// Enum representing trie nodes in sparse trie.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum SparseNode {
-    /// Empty trie node.
-    Empty,
-    /// The hash of the node that was not revealed.
-    Hash(B256),
-    /// Sparse leaf node with remaining key suffix.
-    Leaf {
-        /// Remaining key suffix for the leaf node.
-        key: Nibbles,
-        /// Pre-computed hash of the sparse node.
-        /// Can be reused unless this trie path has been updated.
-        hash: Option<B256>,
-    },
-    /// Sparse extension node with key.
-    Extension {
-        /// The key slice stored by this extension node.
-        key: Nibbles,
-        /// Pre-computed hash of the sparse node.
-        /// Can be reused unless this trie path has been updated.
-        ///
-        /// If [`None`], then the value is not known and should be calculated from scratch.
-        hash: Option<B256>,
-        /// Pre-computed flag indicating whether the trie node should be stored in the database.
-        /// Can be reused unless this trie path has been updated.
-        ///
-        /// If [`None`], then the value is not known and should be calculated from scratch.
-        store_in_db_trie: Option<bool>,
-    },
-    /// Sparse branch node with state mask.
-    Branch {
-        /// The bitmask representing children present in the branch node.
-        state_mask: TrieMask,
-        /// Pre-computed hash of the sparse node.
-        /// Can be reused unless this trie path has been updated.
-        ///
-        /// If [`None`], then the value is not known and should be calculated from scratch.
-        hash: Option<B256>,
-        /// Pre-computed flag indicating whether the trie node should be stored in the database.
-        /// Can be reused unless this trie path has been updated.
-        ///
-        /// If [`None`], then the value is not known and should be calculated from scratch.
-        store_in_db_trie: Option<bool>,
-    },
-}
-
-impl SparseNode {
-    /// Create new sparse node from [`TrieNode`].
-    pub fn from_node(node: TrieNode) -> Self {
-        match node {
-            TrieNode::EmptyRoot => Self::Empty,
-            TrieNode::Leaf(leaf) => Self::new_leaf(leaf.key),
-            TrieNode::Extension(ext) => Self::new_ext(ext.key),
-            TrieNode::Branch(branch) => Self::new_branch(branch.state_mask),
-        }
-    }
-
-    /// Create new [`SparseNode::Branch`] from state mask.
-    pub const fn new_branch(state_mask: TrieMask) -> Self {
-        Self::Branch { state_mask, hash: None, store_in_db_trie: None }
-    }
-
-    /// Create new [`SparseNode::Branch`] with two bits set.
-    pub const fn new_split_branch(bit_a: u8, bit_b: u8) -> Self {
-        let state_mask = TrieMask::new(
-            // set bits for both children
-            (1u16 << bit_a) | (1u16 << bit_b),
-        );
-        Self::Branch { state_mask, hash: None, store_in_db_trie: None }
-    }
-
-    /// Create new [`SparseNode::Extension`] from the key slice.
-    pub const fn new_ext(key: Nibbles) -> Self {
-        Self::Extension { key, hash: None, store_in_db_trie: None }
-    }
-
-    /// Create new [`SparseNode::Leaf`] from leaf key and value.
-    pub const fn new_leaf(key: Nibbles) -> Self {
-        Self::Leaf { key, hash: None }
-    }
-
-    /// Returns `true` if the node is a hash node.
-    pub const fn is_hash(&self) -> bool {
-        matches!(self, Self::Hash(_))
-    }
-}
-
-/// A helper struct used to store information about a node that has been removed
-/// during a deletion operation.
-#[derive(Debug)]
-struct RemovedSparseNode {
-    /// The path at which the node was located.
-    path: Nibbles,
-    /// The removed node
-    node: SparseNode,
-    /// For branch nodes, an optional nibble that should be unset due to the node being removed.
-    ///
-    /// During leaf deletion, this identifies the specific branch nibble path that
-    /// connects to the leaf being deleted. Then when restructuring the trie after deletion,
-    /// this nibble position will be cleared from the branch node's to
-    /// indicate that the child no longer exists.
-    ///
-    /// This is only set for branch nodes that have a direct path to the leaf being deleted.
-    unset_branch_nibble: Option<u8>,
-}
-
-/// Collection of reusable buffers for [`RevealedSparseTrie::rlp_node`] calculations.
-///
-/// These buffers reduce allocations when computing RLP representations during trie updates.
-#[derive(Debug, Default)]
-pub struct RlpNodeBuffers {
-    /// Stack of RLP node paths
-    path_stack: Vec<RlpNodePathStackItem>,
-    /// Stack of RLP nodes
-    rlp_node_stack: Vec<RlpNodeStackItem>,
-    /// Reusable branch child path
-    branch_child_buf: SmallVec<[Nibbles; 16]>,
-    /// Reusable branch value stack
-    branch_value_stack_buf: SmallVec<[RlpNode; 16]>,
-}
-
-impl RlpNodeBuffers {
-    /// Creates a new instance of buffers with the root path on the stack.
-    fn new_with_root_path() -> Self {
-        Self {
-            path_stack: vec![RlpNodePathStackItem {
-                level: 0,
-                path: Nibbles::default(),
-                is_in_prefix_set: None,
-            }],
-            rlp_node_stack: Vec::new(),
-            branch_child_buf: SmallVec::<[Nibbles; 16]>::new_const(),
-            branch_value_stack_buf: SmallVec::<[RlpNode; 16]>::new_const(),
-        }
-    }
-}
-
-/// RLP node path stack item.
-#[derive(Debug)]
-struct RlpNodePathStackItem {
-    /// Level at which the node is located. Higher numbers correspond to lower levels in the trie.
-    level: usize,
-    /// Path to the node.
-    path: Nibbles,
-    /// Whether the path is in the prefix set. If [`None`], then unknown yet.
-    is_in_prefix_set: Option<bool>,
-}
-
-/// RLP node stack item.
-#[derive(Debug)]
-struct RlpNodeStackItem {
-    /// Path to the node.
-    path: Nibbles,
-    /// RLP node.
-    rlp_node: RlpNode,
-    /// Type of the node.
-    node_type: SparseNodeType,
-}
-
-/// Tracks modifications to the sparse trie structure.
-///
-/// Maintains references to both modified and pruned/removed branches, enabling
-/// one to make batch updates to a persistent database.
-#[derive(Debug, Clone, Default, PartialEq, Eq)]
-pub struct SparseTrieUpdates {
-    pub(crate) updated_nodes: HashMap<Nibbles, BranchNodeCompact>,
-    pub(crate) removed_nodes: HashSet<Nibbles>,
-    pub(crate) wiped: bool,
-}
-
-impl SparseTrieUpdates {
-    /// Create new wiped sparse trie updates.
-    pub fn wiped() -> Self {
-        Self { wiped: true, ..Default::default() }
-    }
-
-    /// Clears the updates, but keeps the backing data structures allocated.
-    ///
-    /// Sets `wiped` to `false`.
-    pub fn clear(&mut self) {
-        self.updated_nodes.clear();
-        self.removed_nodes.clear();
-        self.wiped = false;
     }
 }
